Combine AC and React mock link and subscription testing utils
