+++ src/__tests__/graphqlSubscriptions.ts
@@ -1,6 +1,9 @@
 import gql from 'graphql-tag';
 
-import { mockObservableLink, MockedSubscription } from './utils/mocks/mockLinks';
+import {
+  mockObservableLink,
+  MockedSubscription
+} from './utils/mocks/mockSubscriptionLink';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
 import { ApolloClient } from '../';
 import { QueryManager } from '../core/QueryManager';
