+++ src/core/__tests__/QueryManager/recycler.ts
@@ -8,7 +8,9 @@
 import gql from 'graphql-tag';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
 import { stripSymbols } from '../../../__tests__/utils/stripSymbols';
-import { MockSubscriptionLink } from '../../../__tests__/utils/mocks/mockLinks';
+import {
+  MockSubscriptionLink
+} from '../../../__tests__/utils/mocks/mockSubscriptionLink';
 
 // core
 import { QueryManager } from '../../QueryManager';
