+++ src/react/testing/index.ts
@@ -1,12 +1,10 @@
 export { MockedProvider } from './mocks/MockedProvider';
-export { MockLink, mockSingleLink } from './mocks/mockLink';
+export { MockLink, mockSingleLink } from '../../__tests__/utils/mocks/mockLink';
 export {
   MockSubscriptionLink,
   mockObservableLink
-} from './mocks/mockSubscriptionLink';
+} from '../../__tests__/utils/mocks/mockSubscriptionLink';
 
 export { createClient } from './utils/createClient';
 export { stripSymbols } from '../../__tests__/utils/stripSymbols';
 export { wait } from './utils/wait';
-
-export * from './mocks/types';
